Implement cli `destroy` command using zenoh-based control channel
