+++ binaries/coordinator/src/lib.rs
@@ -1,17 +1,11 @@
-use dora_core::descriptor::{self, collect_dora_timers, CoreNodeKind, Descriptor};
-use dora_node_api::{
-    communication,
-    config::{format_duration, NodeId},
-};
-use eyre::{bail, eyre, WrapErr};
-use futures::{stream::FuturesUnordered, FutureExt, StreamExt};
+use eyre::WrapErr;
+use futures::StreamExt;
 use futures_concurrency::stream::Merge;
-use std::{
-    env::consts::EXE_EXTENSION,
-    path::{Path, PathBuf},
-    time::Duration,
-};
-use tokio_stream::wrappers::{IntervalStream, ReceiverStream};
+use std::path::{Path, PathBuf};
+use tokio_stream::wrappers::ReceiverStream;
+
+mod control;
+mod run;
 
 #[derive(Debug, Clone, clap::Parser)]
 #[clap(about = "Dora coordinator")]
@@ -37,7 +31,7 @@ pub async fn run(args: Args) -> eyre::Result<()> {
     match run_dataflow {
         Some(path) => {
             // start the given dataflow directly
-            self::run_dataflow(path.clone(), &runtime_path)
+            run::run_dataflow(path.clone(), &runtime_path)
                 .await
                 .wrap_err_with(|| format!("failed to run dataflow at {}", path.display()))?;
         }
@@ -55,16 +49,24 @@ async fn start(runtime_path: &Path) -> eyre::Result<()> {
     let mut dataflow_errors_tx = Some(dataflow_errors_tx);
     let dataflow_error_events = ReceiverStream::new(dataflow_errors).map(Event::DataflowError);
 
-    let stop_events = tokio::time::sleep(Duration::from_secs(5))
-        .into_stream()
-        .map(|()| Event::Stop);
+    let control_events = control::control_events();
 
-    let mut events = (dataflow_error_events, stop_events).merge();
+    let mut events = (dataflow_error_events, control_events).merge();
 
     while let Some(event) = events.next().await {
+        let mut stop = false;
         match event {
             Event::DataflowError(err) => {
                 tracing::error!("{err:?}");
+            }
+            Event::ParseError(err) => {
+                let err = err.wrap_err("failed to parse message");
+                tracing::error!("{err:?}");
+            }
+            Event::ControlChannelError(err) => {
+                let err = err.wrap_err("Stopping because of fatal control channel error");
+                tracing::error!("{err:?}");
+                stop = true;
             }
             Event::StartDataflow { path } => {
                 let runtime_path = runtime_path.to_owned();
@@ -76,7 +78,7 @@ async fn start(runtime_path: &Path) -> eyre::Result<()> {
                     }
                 };
                 let task = async move {
-                    let result = run_dataflow(path.clone(), &runtime_path)
+                    let result = run::run_dataflow(path.clone(), &runtime_path)
                         .await
                         .wrap_err_with(|| format!("failed to run dataflow at {}", path.display()));
                     match result {
@@ -90,216 +92,27 @@ async fn start(runtime_path: &Path) -> eyre::Result<()> {
             }
             Event::Stop => {
                 tracing::info!("Received stop command");
-                // ensure that no new dataflows can be started
-                dataflow_errors_tx = None;
+                stop = true;
             }
         }
+
+        if stop {
+            tracing::info!("stopping...");
+
+            // ensure that no new dataflows can be started
+            dataflow_errors_tx = None;
+        }
     }
+
+    tracing::info!("stopped");
 
     Ok(())
 }
 
 enum Event {
     DataflowError(eyre::Report),
+    ControlChannelError(eyre::Report),
     StartDataflow { path: PathBuf },
+    ParseError(eyre::Report),
     Stop,
-}
-
-async fn run_dataflow(dataflow_path: PathBuf, runtime: &Path) -> eyre::Result<()> {
-    let runtime = runtime.with_extension(EXE_EXTENSION);
-    let descriptor = read_descriptor(&dataflow_path).await.wrap_err_with(|| {
-        format!(
-            "failed to read dataflow descriptor at {}",
-            dataflow_path.display()
-        )
-    })?;
-
-    let nodes = descriptor.resolve_aliases();
-    let dora_timers = collect_dora_timers(&nodes);
-    let communication_config = {
-        let mut config = descriptor.communication;
-        // add uuid as prefix to ensure isolation
-        config.add_topic_prefix(&uuid::Uuid::new_v4().to_string());
-        config
-    };
-
-    if nodes
-        .iter()
-        .any(|n| matches!(n.kind, CoreNodeKind::Runtime(_)))
-        && !runtime.is_file()
-    {
-        bail!(
-            "There is no runtime at {}, or it is not a file",
-            runtime.display()
-        );
-    }
-
-    let mut tasks = FuturesUnordered::new();
-    for node in nodes {
-        let node_id = node.id.clone();
-
-        match node.kind {
-            descriptor::CoreNodeKind::Custom(node) => {
-                let result = spawn_custom_node(node_id.clone(), &node, &communication_config)
-                    .wrap_err_with(|| format!("failed to spawn custom node {node_id}"))?;
-                tasks.push(result);
-            }
-            descriptor::CoreNodeKind::Runtime(node) => {
-                if !node.operators.is_empty() {
-                    let result =
-                        spawn_runtime_node(&runtime, node_id.clone(), &node, &communication_config)
-                            .wrap_err_with(|| format!("failed to spawn runtime node {node_id}"))?;
-                    tasks.push(result);
-                }
-            }
-        }
-    }
-
-    for interval in dora_timers {
-        let communication_config = communication_config.clone();
-        let mut communication =
-            tokio::task::spawn_blocking(move || communication::init(&communication_config))
-                .await
-                .wrap_err("failed to join communication layer init task")?
-                .wrap_err("failed to init communication layer")?;
-        tokio::spawn(async move {
-            let topic = {
-                let duration = format_duration(interval);
-                format!("dora/timer/{duration}")
-            };
-            let metadata = dora_message::Metadata::default();
-            let data = metadata.serialize().unwrap();
-            let mut stream = IntervalStream::new(tokio::time::interval(interval));
-            while (stream.next().await).is_some() {
-                communication
-                    .publisher(&topic)
-                    .unwrap()
-                    .publish(&data)
-                    .expect("failed to publish timer tick message");
-            }
-        });
-    }
-
-    while let Some(task_result) = tasks.next().await {
-        task_result
-            .wrap_err("failed to join async task")?
-            .wrap_err("custom node failed")?;
-    }
-
-    Ok(())
-}
-
-#[tracing::instrument]
-fn spawn_custom_node(
-    node_id: NodeId,
-    node: &descriptor::CustomNode,
-    communication: &dora_node_api::config::CommunicationConfig,
-) -> eyre::Result<tokio::task::JoinHandle<eyre::Result<(), eyre::Error>>> {
-    let mut args = node.run.split_ascii_whitespace();
-    let cmd = {
-        let raw = Path::new(
-            args.next()
-                .ok_or_else(|| eyre!("`run` field must not be empty"))?,
-        );
-        let path = if raw.extension().is_none() {
-            raw.with_extension(EXE_EXTENSION)
-        } else {
-            raw.to_owned()
-        };
-        path.canonicalize()
-            .wrap_err_with(|| format!("no node exists at `{}`", path.display()))?
-    };
-
-    let mut command = tokio::process::Command::new(cmd);
-    command.args(args);
-    command_init_common_env(&mut command, &node_id, communication)?;
-    command.env(
-        "DORA_NODE_RUN_CONFIG",
-        serde_yaml::to_string(&node.run_config)
-            .wrap_err("failed to serialize custom node run config")?,
-    );
-
-    // Injecting the env variable defined in the `yaml` into
-    // the node runtime.
-    if let Some(envs) = &node.env {
-        for (key, value) in envs {
-            command.env(key, value.to_string());
-        }
-    }
-
-    let mut child = command
-        .spawn()
-        .wrap_err_with(|| format!("failed to run command `{}`", &node.run))?;
-    let result = tokio::spawn(async move {
-        let status = child.wait().await.context("child process failed")?;
-        if status.success() {
-            tracing::info!("node {node_id} finished");
-            Ok(())
-        } else if let Some(code) = status.code() {
-            Err(eyre!("node {node_id} failed with exit code: {code}"))
-        } else {
-            Err(eyre!("node {node_id} failed (unknown exit code)"))
-        }
-    });
-    Ok(result)
-}
-
-#[tracing::instrument(skip(node))]
-fn spawn_runtime_node(
-    runtime: &Path,
-    node_id: NodeId,
-    node: &descriptor::RuntimeNode,
-    communication: &dora_node_api::config::CommunicationConfig,
-) -> eyre::Result<tokio::task::JoinHandle<eyre::Result<(), eyre::Error>>> {
-    let mut command = tokio::process::Command::new(runtime);
-    command_init_common_env(&mut command, &node_id, communication)?;
-    command.env(
-        "DORA_OPERATORS",
-        serde_yaml::to_string(&node.operators)
-            .wrap_err("failed to serialize custom node run config")?,
-    );
-
-    let mut child = command
-        .spawn()
-        .wrap_err_with(|| format!("failed to run runtime at `{}`", runtime.display()))?;
-    let result = tokio::spawn(async move {
-        let status = child.wait().await.context("child process failed")?;
-        if status.success() {
-            tracing::info!("runtime node {node_id} finished");
-            Ok(())
-        } else if let Some(code) = status.code() {
-            Err(eyre!(
-                "runtime node {node_id} failed with exit code: {code}"
-            ))
-        } else {
-            Err(eyre!("runtime node {node_id} failed (unknown exit code)"))
-        }
-    });
-    Ok(result)
-}
-
-fn command_init_common_env(
-    command: &mut tokio::process::Command,
-    node_id: &NodeId,
-    communication: &dora_node_api::config::CommunicationConfig,
-) -> Result<(), eyre::Error> {
-    command.env(
-        "DORA_NODE_ID",
-        serde_yaml::to_string(&node_id).wrap_err("failed to serialize custom node ID")?,
-    );
-    command.env(
-        "DORA_COMMUNICATION_CONFIG",
-        serde_yaml::to_string(communication)
-            .wrap_err("failed to serialize communication config")?,
-    );
-    Ok(())
-}
-
-async fn read_descriptor(file: &Path) -> Result<Descriptor, eyre::Error> {
-    let descriptor_file = tokio::fs::read(file)
-        .await
-        .context("failed to open given file")?;
-    let descriptor: Descriptor =
-        serde_yaml::from_slice(&descriptor_file).context("failed to parse given descriptor")?;
-    Ok(descriptor)
 }
