+++ libraries/core/src/topics.rs
@@ -1,0 +1,3 @@
+pub const ZENOH_CONTROL_PREFIX: &str = "dora_control";
+pub const ZENOH_CONTROL_STOP_ALL: &str = "stop_all";
+pub const ZENOH_CONTROL_START_DATAFLOW: &str = "start_dataflow";
